test/requirements.txt
pandas==1.0.5
psycopg2-binary==2.8.5
SQLAlchemy==1.3.18